requirements.txt
# Packages that should be installed on ubuntu before requirements install
# libmysqlclient-dev libpq-dev libsqlite3-dev libjpeg-dev libfreetype6 libfreetype6-dev zlib1g-dev

# python-dev (for python environment) | pypy-dev (for pypy-environment)


# for deployment
# gunicorn

# for python2.7
psycopg2

# for pypy
# psycopg2cffi


Django
Jinja2
MySQL-python
Pillow==2.5
celery
hiredis
mandrill
mmh3
phonenumbers
pysqlite
pytz
redis
httplib2
google-api-python-client
oauth2client
pyOpenSSL
billiard
python-redis-lock
kombu
django_compressor
django-redis-cache
djorm-pgarray
amqp
anyjson
six
mock
nose
#rjsmin # js minifier
